+++ src/Controllers/UserData.vala
@@ -23,6 +23,7 @@ using ThiefMD.Widgets;
 namespace ThiefMD.Controllers.UserData {
     public string data_path;
     public string style_path;
+
     public void create_data_directories () {
         data_path = Path.build_path (
                         Path.DIR_SEPARATOR_S,
@@ -42,5 +43,29 @@ namespace ThiefMD.Controllers.UserData {
         } catch (Error e) {
             warning ("Error: %s\n", e.message);
         }
+    }
+
+    public string? get_trash_folder () {
+        string? home_folder = Environment.get_variable ("XDG_DATA_HOME");
+        string? trash_folder = null;
+        if (home_folder == null) {
+            home_folder = Environment.get_variable ("HOME");
+            if (home_folder == null) {
+                home_folder = Environment.get_home_dir ();
+            }
+            trash_folder = Path.build_path (
+                                Path.DIR_SEPARATOR_S,
+                                home_folder,
+                                ".local",
+                                "share",
+                                "Trash");
+        } else {
+            trash_folder = Path.build_path (
+                Path.DIR_SEPARATOR_S,
+                home_folder,
+                "Trash");
+        }
+        warning ("Using %s for trash", trash_folder);
+        return trash_folder;
     }
 }