+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2020-08-29
 
  * Add ability to hide folders in the Library. This will create or modify a .thiefignore file to the directory.
+ * Add ability to move sheets to Trash from Sheets view.
 
 ## Prior